Add refactor each code into folders
