+++ rental/rentals_handler.rb
